Fix -i option handling.
ok reyk@

+++ usr.sbin/vmmctl/parser.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: parser.h,v 1.3 2015/11/27 09:11:39 reyk Exp $	*/
+/*	$OpenBSD: parser.h,v 1.4 2015/12/01 20:52:44 halex Exp $	*/
 
 /*
  * Copyright (c) 2015 Reyk Floeter <reyk@openbsd.org>
@@ -41,7 +41,7 @@ struct parse_result {
 	char			*name;
 	char			*path;
 	long long		 size;
-	size_t			 nifs;
+	int			 nifs;
 	size_t			 ndisks;
 	char			**disks;
 	int			 disable;
